+++ app/Http/Middleware/HandleInertiaRequests.php
@@ -78,6 +78,7 @@ class HandleInertiaRequests extends Middleware
             'flash' => [
                 'success' => $request->session()->get('success'),
                 'error' => $request->session()->get('error'),
+                'warning' => $request->session()->get('warning'),
             ],
             'packages' => (new Module())->allModules(),
             'adminAllSetting' =>   $request->user() ?  getAdminAllSetting() : getAdminAllSetting(true),